Refactored neovim code
