+++ extension.ts
@@ -20,6 +20,8 @@ import { ICodeKeybinding } from './src/mode/remapper';
 import { runCmdLine } from './src/cmd_line/main';
 
 import './src/actions/vim.all';
+import { attach } from "promised-neovim-client";
+import { spawn } from "child_process";
 
 interface VSCodeKeybinding {
   key: string;
@@ -81,6 +83,8 @@ export async function getAndUpdateModeHandler(): Promise<ModeHandler> {
   let curHandler = modeHandlerToEditorIdentity[activeEditorId.toString()];
   if (!curHandler) {
     const newModeHandler = new ModeHandler();
+    const proc = spawn('nvim', ['-u', 'NONE', '-N', '--embed'], {cwd: __dirname });
+    newModeHandler.vimState.nvim = await attach(proc.stdin, proc.stdout);
 
     modeHandlerToEditorIdentity[activeEditorId.toString()] = newModeHandler;
     extensionContext.subscriptions.push(newModeHandler);
